requirements.txt
sounddevice
numpy
pyaudio
daily-python
vapi_server_sdk
python-dotenv
gpiozero
requests
